Fix autocomplete on search results page (IE make it appear again)
In ecad06e390e66a6b7a16ee624df3de9121410c04v we use a selector that targets the mobile autocomplete on all pages except our search results page, because of how google assigns classes (see inline comment in gcse-search.css).  So we need to make a distinction between the autocomplete selector that we use on search results page vs all other pages to account for this.  This distinction is facilitated by a body class which we use js to add in our search results custom js.