Preprocess the emergency alert node to populate data model for template
Based on the emergency alert docs in mayflower (@organisms/by-template/emergency-alerts.json) we know the data model that is required for the pattern template.  We use the preprocess function to route Drupal's data to the data model and make it available in the emergency alert node template for the teaser display mode, which we use for the alert banner.