Make emergency alert banner button show/hide preference persist
We pass the node updated timestamp to the data object which populates a data attribute in the alert banner html container element.  A session cookie checks for the status of the session cookie (hide/show) and so long as there have been no updates to the content, the state of the banner is respected.